backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
python-dotenv